Changed Exponential to Long and updated resources to make sense

+++ core/src/com/blackphoton/planetclicker/objectType/table/entries/TableEntry.java
@@ -84,7 +84,7 @@ public abstract class TableEntry {
 		if(createClicked){
 			create.setDrawable(new SpriteDrawable(new Sprite(createClicked_tex)));
 		}else
-		if(Data.main.getPopulationCount()>=requiredEra.getPop_req().toInt()){
+		if(Data.main.getPopulationCount()>=requiredEra.getPop_req()){
 			create.setDrawable(new SpriteDrawable(new Sprite(create_tex)));
 			canCreate = true;
 		}else{
@@ -99,7 +99,7 @@ public abstract class TableEntry {
 			upgrade.setDrawable(new SpriteDrawable(new Sprite(upgradeLocked_tex)));
 			canUpgrade = false;
 		}else
-		if(Data.main.getPopulationCount()>=upgradeTo.requiredEra.getPop_req().toInt()){
+		if(Data.main.getPopulationCount()>=upgradeTo.requiredEra.getPop_req()){
 			upgrade.setDrawable(new SpriteDrawable(new Sprite(upgrade_tex)));
 			canUpgrade = true;
 		}else{
